// app/entities/header-data.interface.ts
export interface HeaderData {
    linkLogoRentas: string;
    loginStatusOptions: {
        linkCidiLogin: string;
        linkCidiSignIn: string;
        linkAfipLogin: string;
        linkAfipSingIn: string;
        linkMiPerfil: string;
        linkFiscalizaciones: string;
        linkCorredor: string;
        linkMartillero: string;
        linkProcurador: string;
        linkMisRepresentados: string;
        linkLogOut: string;
    };
    linkTramite: string;
    linkConstanciaRegular: string;
}